Added all current files
make_df.ipynb, is the only code file in this repository so far. it takes in a .mat file from (.    ) and it reads in the raw current / amplitude data. it then makes a dataframe for for that file, which contains the baseline amplitude found from the baseline sweeps, the drug used on each sweep, the normalized amplitude relative to the baseline. and then uses this data to make different plots. you need to supply specific information at the start to get the code to run.